detangle: Remove dependency on watcher from extensions
